feat: replace employee signup page with a login/register card (#293)
Combines student login and employee registration into one dark card on
/login with an "Estudante" / "Login" tab split by authentication method:
AuthNEI (student-only, login and signup in one OAuth click) versus the
password form used by admins and returning employees. Employee
registration hangs off the Login tab as a secondary action rather than a
peer tab, so a returning employee's login isn't buried under "Estudante".
The old standalone /signup/employee route now redirects into this card.
